update (OSIReporter): OSI timings reverted again. Something interesting going on at this line

+++ EnvironmentSimulator/Modules/ScenarioEngine/SourceFiles/OSIReporter.cpp
@@ -382,7 +382,7 @@ int OSIReporter::UpdateOSIDynamicGroundTruth(const std::vector<std::unique_ptr<O
 	{
 		// use timstamp from object state
 		obj_osi_internal.gt->mutable_timestamp()->set_seconds(static_cast<int64_t>(objectState[0]->state_.info.timeStamp));
-		obj_osi_internal.gt->mutable_timestamp()->set_nanos(static_cast<uint32_t>(static_cast<int64_t>(objectState[0]->state_.info.timeStamp) - static_cast<int64_t>(objectState[0]->state_.info.timeStamp * 1e9)));
+		obj_osi_internal.gt->mutable_timestamp()->set_nanos(static_cast<uint32_t>(((objectState[0]->state_.info.timeStamp - static_cast<int64_t>(objectState[0]->state_.info.timeStamp)) * 1e9)));
 	}
 	else
 	{
